+++ testData/org/elixir_lang/heex/file/eex_data/page_live.ex
@@ -1,0 +1,5 @@
+defmodule MyAppWeb.PageLive do
+  def button(assigns) do
+    ~H""
+  end
+end